+++ R/utils_cli.R
@@ -50,7 +50,7 @@ cli_cancel <- function(
   cli_yellow_box(msg, print = print)
 }
 
-cli_uptodate <- function(
+cli_pipeline_uptodate <- function(
   time_stamp = FALSE,
   seconds_elapsed = NULL,
   print = TRUE
@@ -64,7 +64,7 @@ cli_uptodate <- function(
   cli_green_check(msg, print = print)
 }
 
-cli_done <- function(
+cli_pipeline_done <- function(
   time_stamp = FALSE,
   seconds_elapsed = NULL,
   print = TRUE
@@ -78,13 +78,27 @@ cli_done <- function(
   cli_blue_play(msg, print = print)
 }
 
-cli_empty <- function(
+cli_pipeline_empty <- function(
   time_stamp = FALSE,
   seconds_elapsed = NULL,
   print = TRUE
 ) {
   time <- if_any(time_stamp, time_stamp(), NULL)
-  msg <- paste(c(time, "no targets found"), collapse = " ")
+  msg <- paste(c(time, "empty pipeline"), collapse = " ")
+  if (!is.null(seconds_elapsed)) {
+    msg_time <- paste0(" [", units_seconds(seconds_elapsed), "]")
+    msg <- paste0(msg, msg_time)
+  }
+  cli_red_x(msg, print = print)
+}
+
+cli_pipeline_errored <- function(
+  time_stamp = FALSE,
+  seconds_elapsed = NULL,
+  print = TRUE
+) {
+  time <- if_any(time_stamp, time_stamp(), NULL)
+  msg <- paste(c(time, "errored pipeline"), collapse = " ")
   if (!is.null(seconds_elapsed)) {
     msg_time <- paste0(" [", units_seconds(seconds_elapsed), "]")
     msg <- paste0(msg, msg_time)