+++ tests/testthat/test-utils_cli.R
@@ -23,19 +23,24 @@ tar_test("cli_cancel()", {
   expect_message(cli_cancel("x", time_stamp = TRUE))
 })
 
-tar_test("cli_uptodate()", {
+tar_test("cli_pipeline_uptodate()", {
   skip_cran()
-  expect_message(cli_uptodate(seconds_elapsed = 1))
+  expect_message(cli_pipeline_uptodate(seconds_elapsed = 1))
 })
 
-tar_test("cli_done()", {
+tar_test("cli_pipeline_done()", {
   skip_cran()
-  expect_message(cli_done(seconds_elapsed = 1))
+  expect_message(cli_pipeline_done(seconds_elapsed = 1))
 })
 
-tar_test("cli_empty()", {
+tar_test("cli_pipeline_empty()", {
   skip_cran()
-  expect_message(cli_empty(seconds_elapsed = 1))
+  expect_message(cli_pipeline_empty(seconds_elapsed = 1))
+})
+
+tar_test("cli_pipeline_errored()", {
+  skip_cran()
+  expect_message(cli_pipeline_errored(seconds_elapsed = 1))
 })
 
 tar_test("cli_workspace()", {